back-end/requirements.txt
fastapi==0.121.0
psycopg2==2.9.11
python-dotenv==1.2.1
uvicorn==0.38.0